feat(roles-permissions): implement role and permission management with seeding and authorization

+++ routes/admin.php
@@ -11,7 +11,7 @@ use App\Http\Controllers\Admin\PermissionController;
 
 Route::prefix('admin')->group(function () {
 
-    Route::middleware('guest:admin')->group(function () {
+    Route::middleware("guest:admin")->group(function () {
         Route::get('/register', [RegisterController::class, 'showRegistrationForm'])->name('admin.register.form');
         Route::post('/register', [RegisterController::class, 'register'])->name('admin.register');
 
@@ -23,50 +23,50 @@ Route::prefix('admin')->group(function () {
         Route::get('/dashboard', [DashboardController::class, 'index'])->name('admin.dashboard');
 
         Route::prefix('tasks')->group(function () {
-            Route::get('/', [TaskController::class, 'index'])->name('tasks.index');
-            Route::get('/create', [TaskController::class, 'create'])->name('tasks.create');
-            Route::post('/create', [TaskController::class, 'store'])->name('tasks.store');
-            Route::get('/{task}', [TaskController::class, 'show'])->name('tasks.show');
-            Route::get('/{task}/edit', [TaskController::class, 'edit'])->name('tasks.edit');
-            Route::patch('/{task}/update', [TaskController::class, 'update'])->name('tasks.update');
-            Route::delete('/{task}/delete', [TaskController::class, 'destroy'])->name('tasks.destroy');
+            Route::get('/', [TaskController::class, 'index'])->name('tasks.index')->can('manage-tasks');
+            Route::get('/create', [TaskController::class, 'create'])->name('tasks.create')->can('manage-tasks');
+            Route::post('/create', [TaskController::class, 'store'])->name('tasks.store')->can('manage-tasks');
+            Route::get('/{task}', [TaskController::class, 'show'])->name('tasks.show')->can('manage-tasks');
+            Route::get('/{task}/edit', [TaskController::class, 'edit'])->name('tasks.edit')->can('manage-tasks');
+            Route::patch('/{task}/update', [TaskController::class, 'update'])->name('tasks.update')->can('manage-tasks');
+            Route::delete('/{task}/delete', [TaskController::class, 'destroy'])->name('tasks.destroy')->can('manage-tasks');
         });
 
         Route::prefix('interns')->group(function () {
-            Route::get('/', [InternController::class, 'index'])->name('interns.index');
-            Route::get('/create', [InternController::class, 'create'])->name('interns.create');
-            Route::post('/create', [InternController::class, 'store'])->name('interns.store');
-            Route::get('/assign', [InternController::class, 'assign'])->name('interns.assign');
-            Route::post('/assign', [InternController::class, 'assignStore'])->name('interns.assign.store');
-            Route::get('/{iterns}', [InternController::class, 'show'])->name('interns.show');
-            Route::get('/{intern}/edit', [InternController::class, 'edit'])->name('interns.edit');
-            Route::patch('/{intern}/update', [InternController::class, 'update'])->name('interns.update');
-            Route::delete('/{intern}/delete', [InternController::class, 'destroy'])->name('interns.destroy');
+            Route::get('/', [InternController::class, 'index'])->name('interns.index')->can('manage-interns');
+            Route::get('/create', [InternController::class, 'create'])->name('interns.create')->can('manage-interns');
+            Route::post('/create', [InternController::class, 'store'])->name('interns.store')->can('manage-interns');
+            Route::get('/assign', [InternController::class, 'assign'])->name('interns.assign')->can('manage-interns');
+            Route::post('/assign', [InternController::class, 'assignStore'])->name('interns.assign.store')->can('manage-interns');
+            Route::get('/{iterns}', [InternController::class, 'show'])->name('interns.show')->can('manage-interns');
+            Route::get('/{intern}/edit', [InternController::class, 'edit'])->name('interns.edit')->can('manage-interns');
+            Route::patch('/{intern}/update', [InternController::class, 'update'])->name('interns.update')->can('manage-interns');
+            Route::delete('/{intern}/delete', [InternController::class, 'destroy'])->name('interns.destroy')->can('manage-interns');
         });
 
         Route::prefix('admins')->group(function () {
-            Route::get('/', [AdminController::class, 'index'])->name('admins.index');
-            Route::get('/create', [AdminController::class, 'create'])->name('admins.create');
-            Route::post('/create', [AdminController::class, 'store'])->name('admins.store');
-            Route::get('/{admin}/edit', [AdminController::class, 'edit'])->name('admins.edit');
-            Route::patch('/{admin}/update', [AdminController::class, 'update'])->name('admins.update');
-            Route::delete('/{admin}/delete', [AdminController::class, 'destroy'])->name('admins.destroy');
+            Route::get('/', [AdminController::class, 'index'])->name('admins.index')->can('manage-admins');
+            Route::get('/create', [AdminController::class, 'create'])->name('admins.create')->can('manage-admins');
+            Route::post('/create', [AdminController::class, 'store'])->name('admins.store')->can('manage-admins');
+            Route::get('/{admin}/edit', [AdminController::class, 'edit'])->name('admins.edit')->can('manage-admins');
+            Route::patch('/{admin}/update', [AdminController::class, 'update'])->name('admins.update')->can('manage-admins');
+            Route::delete('/{admin}/delete', [AdminController::class, 'destroy'])->name('admins.destroy')->can('manage-admins');
         });
         Route::prefix('roles')->group(function () {
-            Route::get('/', [RoleController::class, 'index'])->name('roles.index');
-            Route::get('/create', [RoleController::class, 'create'])->name('roles.create');
-            Route::post('/create', [RoleController::class, 'store'])->name('roles.store');
-            Route::get('/{role}/edit', [RoleController::class, 'edit'])->name('roles.edit');
-            Route::patch('/{role}/update', [RoleController::class, 'update'])->name('roles.update');
-            Route::delete('/{role}/delete', [RoleController::class, 'destroy'])->name('roles.destroy');
+            Route::get('/', [RoleController::class, 'index'])->name('roles.index')->can('manage-roles');
+            Route::get('/create', [RoleController::class, 'create'])->name('roles.create')->can('manage-roles');
+            Route::post('/create', [RoleController::class, 'store'])->name('roles.store')->can('manage-roles');
+            Route::get('/{role}/edit', [RoleController::class, 'edit'])->name('roles.edit')->can('manage-roles');
+            Route::patch('/{role}/update', [RoleController::class, 'update'])->name('roles.update')->can('manage-roles');
+            Route::delete('/{role}/delete', [RoleController::class, 'destroy'])->name('roles.destroy')->can('manage-roles');
         });
         Route::prefix('permissions')->group(function () {
-            Route::get('/', [PermissionController::class, 'index'])->name('permissions.index');
-            Route::get('/create', [PermissionController::class, 'create'])->name('permissions.create');
-            Route::post('/create', [PermissionController::class, 'store'])->name('permissions.store');
-            Route::get('/{permission}/edit', [PermissionController::class, 'edit'])->name('permissions.edit');
-            Route::patch('/{permission}/update', [PermissionController::class, 'update'])->name('permissions.update');
-            Route::delete('/{permission}/delete', [PermissionController::class, 'destroy'])->name('permissions.destroy');
+            Route::get('/', [PermissionController::class, 'index'])->name('permissions.index')->can('manage-permissions');
+            Route::get('/create', [PermissionController::class, 'create'])->name('permissions.create')->can('manage-permissions');
+            Route::post('/create', [PermissionController::class, 'store'])->name('permissions.store')->can('manage-permissions');
+            Route::get('/{permission}/edit', [PermissionController::class, 'edit'])->name('permissions.edit')->can('manage-permissions');
+            Route::patch('/{permission}/update', [PermissionController::class, 'update'])->name('permissions.update')->can('manage-permissions');
+            Route::delete('/{permission}/delete', [PermissionController::class, 'destroy'])->name('permissions.destroy')->can('manage-permissions');
         });
 
         Route::post('logout', [LoginController::class, 'logout'])->name('admin.logout');
